chore: 마이페이지 구조 변경
